Time-sync config example (#55)
* Create time-sync-logging-dashboard.sh

Script for setting up logging and creating an accuracy dashboard for
accurate time for Compute Engine VMs (GCE version using ops-agent, to
contrast with the GKE version following).

* Create time-sync-vm-creation.sh

Adds scripts for creating a VM with accurate time. Following commit
will demonstrate converting this VM creation flow to GKE daemonset

* Add description of the TimeSync example to README.md

Explain what the time-sync-example folder is demonstrating, to help
our readers' orientation.

* Add a README file for the time-sync example directory

Help our readers understand what this example is demonstrating and how
to use the scripts. Currently only documenting the GCE portion, with
GKE version to follow shortly.

* Add GKE DaemonSet for TimeSync configuration

This commit adds a daemonset which is the counterpart of the compute
engine scripts found in the gce folder. It is creating a GKE cluster
with chrony and PTP-KVM enabled clock synchronization, waiting until
chrony is synchronized before allowing workload to run on the node.

* Add README section about the GKE scripts

Document the added GKE scripts for TimeSync in README.md. Explain the
expected usage and the high level role of each script.

---------

Co-authored-by: Michael Frauenhofer <michaelgeo@google.com>